+++ Questionable/Configuration.cs
@@ -235,6 +235,7 @@ internal sealed class Configuration : IPluginConfiguration
         public string Language { get; set; } = "en";
         public bool HideRemainingTasks { get; set; }
         public bool ClaimMail { get; set; }
+        public bool SameJobThroughoutQuest { get; set; } = true;
     }
 
     internal sealed class StopConfiguration