+++ Questionable/Windows/QuestComponents/EventInfoComponent.cs
@@ -3,6 +3,7 @@ using Dalamud.Interface;
 using Dalamud.Interface.Utility.Raii;
 using Humanizer;
 using Humanizer.Localisation;
+using BannerBg = Lumina.Excel.Sheets.BannerBg;
 using Questionable.Model.Questing;
 namespace Questionable.Windows.QuestComponents;
 
@@ -17,10 +18,11 @@ internal sealed class EventInfoComponent
     Configuration configuration)
 {
     private readonly Configuration _configuration = configuration;
-    private readonly List<EventQuest> _eventQuests =
+    internal static readonly List<EventQuest> EventQuests =
     [
         // Add seasonal events here. If a quest has additional required quests (e.g Make It Rain > Gold Saucer), add a relation in QuestData#L220
-        new(_L("Limited Time Items"), [new UnlockLinkId(568)], DateTime.MaxValue)
+        new(_L("Limited Time Items"), [new UnlockLinkId(568)], DateTime.MaxValue),
+        new(_T<BannerBg>(233), [new QuestId(2141)], AtDailyReset(2026, 10, 5)),
     ];
     private readonly QuestController _questController = questController;
 
@@ -30,13 +32,13 @@ internal sealed class EventInfoComponent
     private readonly QuestTooltipComponent _questTooltipComponent = questTooltipComponent;
     private readonly UiUtils _uiUtils = uiUtils;
 
-    public bool ShouldDraw => _configuration.General.ShowIncompleteSeasonalEvents && _eventQuests.Any(IsIncomplete);
+    public bool ShouldDraw => _configuration.General.ShowIncompleteSeasonalEvents && EventQuests.Any(IsIncomplete);
 
-    private static DateTime AtDailyReset(DateOnly date) => new(date, new(14, 59), DateTimeKind.Utc);
+    private static DateTime AtDailyReset(int year, int month, int day) => new(new(year, month, day), new(14, 59), DateTimeKind.Utc);
 
     public void Draw()
     {
-        foreach (EventQuest eventQuest in _eventQuests)
+        foreach (EventQuest eventQuest in EventQuests)
         {
             if (IsIncomplete(eventQuest))
                 DrawEventQuest(eventQuest);
@@ -106,7 +108,7 @@ internal sealed class EventInfoComponent
 
     public IEnumerable<ElementId> GetCurrentlyActiveEventQuests()
     {
-        return _eventQuests
+        return EventQuests
             .Where(x => x.EndsAtUtc >= DateTime.UtcNow)
             .SelectMany(x => x.QuestIds)
             .Where(ShouldShowQuest);
@@ -118,5 +120,5 @@ internal sealed class EventInfoComponent
                !_questFunctions.IsQuestUnobtainable(elementId);
     }
 
-    private sealed record EventQuest(string Name, List<ElementId> QuestIds, DateTime EndsAtUtc);
+    internal sealed record EventQuest(string Name, List<ElementId> QuestIds, DateTime EndsAtUtc);
 }